Qbackend/requirements.txt
python==3.10.2
django==4.0.2
djangorestframework==3.13.1
npm==8.5.4
node==16.14.0